requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.19.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudevents==1.11.0
cryptography==43.0.3
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dj-database-url==2.2.0
Django==4.2.21
django-auditlog==3.1.2
django-auto-logout==0.5.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-jsonfield==1.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.27.2
drf-yasg==1.21.10
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
Markdown==3.7
MarkupSafe==3.0.2
more-itertools==10.5.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.1
openai==1.52.2
openapi-core==0.19.4
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==25.0
pandas==2.2.3
parse==1.20.2
pathable==0.4.3
pillow==10.4.0
portalocker==2.10.1
prance==23.6.21.0
propcache==0.2.0
psycopg2-binary==2.9.10
pybars4==0.9.13
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.9.0
PyMeta3==0.5.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
semantic-kernel==1.12.1
sentry-sdk==2.28.0
six==1.17.0
sniffio==1.3.1
-e git+https://github.com/createXofficial/solar-system-api.git@278b1df53ed4b73048f44fb60998c45fcd98838c#egg=solarsys_backend
sqlparse==0.5.3
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.3
Werkzeug==3.0.6
whitenoise==6.9.0
wrapt==1.16.0
yarl==1.16.0
zipp==3.20.2
